#11: updating readme, simplifying stuff, and changing file structure
