+++ modes/off_mode/off_page.py
@@ -1,8 +1,8 @@
 from tkinter import Frame
-from page import Page
+from modes.page import Page
 from models.model import Model
 
 
-class Speed(Page):
+class Off(Page):
     def __init__(self, parent: Frame, model: Model) -> None:
-        super().__init__(parent, model, "Speed")
+        super().__init__(parent, model, "Off")
