+++ modes/pit_lane_mode/pit_lane_page.py
@@ -1,9 +1,8 @@
 from tkinter import Frame
-from page import Page
+from modes.page import Page
 from models.model import Model
 
 
-class Off(Page):
+class PitLane(Page):
     def __init__(self, parent: Frame, model: Model) -> None:
-        super().__init__(parent, model, "Off")
-
+        super().__init__(parent, model, "Pit Lane")
