+++ modes/speed_mode/speed_page.py
@@ -1,8 +1,8 @@
 from tkinter import Frame
-from page import Page
+from modes.page import Page
 from models.model import Model
 
 
-class PitLane(Page):
+class Speed(Page):
     def __init__(self, parent: Frame, model: Model) -> None:
-        super().__init__(parent, model, "Pit Lane")
+        super().__init__(parent, model, "Speed")
